feat: enhance Supabase client configuration for environment handling
- Updated Supabase client creation to support different environments (production, preview, development) by introducing a new function to determine the environment based on the project reference.
- Refactored the Supabase client to cache instances for each environment, improving performance and resource management.
- Adjusted the API route for verification to utilize the new environment-aware Supabase client.
